add DOCKERFILE and docker compose
