--- crate_db/api/ga4gh/controllers/__init__.py
@@ -1,1 +1,0 @@
-""" Crate DB GA4GH service controller package."""
