--- crate_db/api/ga4gh/controllers/service_info.py
@@ -1,23 +1,0 @@
-"""Crate DB API controllers."""
-
-import logging
-from foca.utils.logging import log_traffic  # type: ignore
-from crate_db.exceptions import NotImplemented
-
-# Get logger instance
-logger = logging.getLogger(__name__)
-
-
-@log_traffic
-def createServiceInfo(service_info) -> dict:
-    return NotImplemented
-
-
-@log_traffic
-def readServiceInfo() -> dict:
-    return NotImplemented
-
-
-@log_traffic
-def updateServiceInfo(service_info) -> dict:
-    return NotImplemented
