+++ crate_db/api/ga4gh/service_info.py
@@ -1,0 +1,1 @@
+"""Service info for crate_db."""
