+++ docker-compose.yaml
@@ -1,0 +1,19 @@
+version: '3.6'
+services:
+
+  crate-db:
+    build:
+        context: .
+        dockerfile: Dockerfile
+    restart: unless-stopped
+    links: 
+        - mongodb
+    command: bash -c "cd /app/crate_db; python app.py"
+    ports:
+        - "8080:8080"
+
+  mongodb: 
+    image: mongo:3.6
+    restart: unless-stopped
+    ports:
+        - "27017:27017"
